requirements.txt
# Core AI/LLM
langchain>=0.3.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0

# GitHub API
PyGithub>=2.1.0

# CLI & Output
rich>=13.7.0
typer>=0.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
